Fix flaky test on python 2.7
Something is killing the skein daemon between tests on python 2.7 only.
I've determined it's not a bug in:
- Skein (close/del is implemented properly and tested in skein)
- dask-yarn (finalizers aren't being called)
- py.test (daemon persists between tests if dask isn't used)

It may be a weird interaction between py.test and signal handling that
distributed/tornado sets up, idk. Since python 2.7 is going away
shortly, I'm just going to punt on this. For now we just create a new
client for each test and things work fine.