Add tests, fix bug in shutdown
This is a bit of a hack for now. To ensure that the exit code of the
client process mirrors the final status of the yarn application, we need
to add our own exception handler, and monkeypatch sys.exit. This is
because atexit hooks don't get the reason for the shutdown (whether it's
a proper exit, or due to an error). A better way would be to bake this
behavior into skein, but I'm not sure how I want to do this yet.