[WIP] Add `dask-yarn submit` command (#35)
* Add `dask-yarn submit` command

Adds a submit command, which submits a script to run independently on
the cluster.

Still needs some iteration on the api, but functionally already works.

* Shutdown cluster on exit from client process

* Add status and kill cli commands

* Add tests, fix bug in shutdown

This is a bit of a hack for now. To ensure that the exit code of the
client process mirrors the final status of the yarn application, we need
to add our own exception handler, and monkeypatch sys.exit. This is
because atexit hooks don't get the reason for the shutdown (whether it's
a proper exit, or due to an error). A better way would be to bake this
behavior into skein, but I'm not sure how I want to do this yet.

* Note the cli as experimental

* Fixup tests

- Better logging on test failure
- Install from conda-forge for more up to date grpcio without segfaults

* Use a separate process to run the script

This makes shutdown much more robust, as the user can't mess with our
logic here. It's also much simpler.

* Make service commands subcommands

Remove the scheduler, worker, and client commands from normal public
view by putting them one level deeper in the cli. Still available to
expert users, but won't be seen by normal users in the top-level cli
help, avoiding potential confusion.

* Update version requirements

To avoid segfaults for submitted scripts, we require grpcio >= 1.14.0.
Install grpcio and friends from pypi, since the versions on conda are
old.

* fixups