Add deploy_mode
Adds a `deploy_mode` parameter, which specifies whether to run the
scheduler/client locally, or on a remote machine (default). This is
similar to spark's deploy-mode, except we use 'local' and 'remote'
instead of 'client' and 'cluster' respectively.

This parameter is useful for debugging, as it makes it easier to access
the dashboard (since the dashboard is running locally), as well as
easier to swap between having scripts run on a local machine vs remote.