Support temporary security credentials (#45)
Add option to submit applications with temporary security credentials.
In this mode every cluster gets a set of transient security credentials
that only exist as long as the application is running. This makes it
easier to have a single process submit applications for multiple users
without leaking security credentials between users.