Better error on application failure (#72)
We used to let the `skein` errors raise if the application failed to
start. This led to many users filing issues in the skein repository when
their application's failed.

We now catch a subset of errors during application startup and raise a
more informative error that describes debugging tips.

This should hopefully be easier for user's to understand.